refactor: 공유 토글 박스 컴포넌트화
